+++ android/src/main/java/com/reactnativestripesdk/PaymentMethodCreateParamsFactory.kt
@@ -42,6 +42,7 @@ class PaymentMethodCreateParamsFactory(
         PaymentMethod.Type.Bancontact -> createBancontactParams()
         PaymentMethod.Type.Billie -> createBillieParams()
         PaymentMethod.Type.SepaDebit -> createSepaParams()
+        PaymentMethod.Type.Multibanco -> createMultibancoParams()
         PaymentMethod.Type.Oxxo -> createOXXOParams()
         PaymentMethod.Type.Eps -> createEPSParams()
         PaymentMethod.Type.GrabPay -> createGrabPayParams()
@@ -129,6 +130,26 @@ class PaymentMethodCreateParamsFactory(
     }
 
     throw PaymentMethodCreateParamsException("You must provide billing details")
+  }
+
+  @Throws(PaymentMethodCreateParamsException::class)
+  private fun createMultibancoParams(): PaymentMethodCreateParams {
+    billingDetailsParams?.let {
+      if (it.email.isNullOrBlank()) {
+        throw PaymentMethodCreateParamsException(
+          "When creating a Multibanco payment method, you must provide the following billing details: email",
+        )
+      }
+
+      return PaymentMethodCreateParams.createMultibanco(
+        billingDetails = it,
+        metadata = metadataParams,
+      )
+    }
+
+    throw PaymentMethodCreateParamsException(
+      "When creating a Multibanco payment method, you must provide the following billing details: email",
+    )
   }
 
   @Throws(PaymentMethodCreateParamsException::class)
@@ -255,6 +276,7 @@ class PaymentMethodCreateParamsFactory(
         PaymentMethod.Type.Bancontact,
         PaymentMethod.Type.Billie,
         PaymentMethod.Type.SepaDebit,
+        PaymentMethod.Type.Multibanco,
         PaymentMethod.Type.Oxxo,
         PaymentMethod.Type.Eps,
         PaymentMethod.Type.GrabPay,