+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 **Fixes**
 * [Fixed] `autocompleteCountries` prop on `AddressSheet` was silently ignored on iOS.
 
+**Features**
+* [Added] Added Multibanco support on iOS and Android. ([#2352](https://github.com/stripe/stripe-react-native/issues/2352))
+
 ## 0.63.0 - 2026-04-06
 **Changes**
 - Updated Stripe iOS SDK from 25.9.+ to 25.10.+.